// frontend/src/stores/types.ts
export interface Holding {
    shares: number
    price: number
    value: number
    name?: string
    currentPrice?: number
}
// ...
export interface ETFPrice {
    price: number
    name: string
}